requirements.txt
bottle==0.13.2
configparser==7.1.0
mysql-connector-python==9.1.0
